moved tests to own project
