+++ wookiee-grpc/src/main/scala/com/oracle/infy/wookiee/grpc/WookieeGrpcUtils.scala
@@ -1,6 +1,6 @@
 package com.oracle.infy.wookiee.grpc
 
-import cats.effect.IO
+import cats.effect.{IO, Resource}
 import com.oracle.infy.wookiee.grpc.impl.GRPCUtils.curatorFramework
 import org.apache.curator.framework.CuratorFramework
 import org.apache.curator.retry.RetryForever
@@ -14,13 +14,17 @@ object WookieeGrpcUtils {
       zkQuorumString: String,
       retryInterval: FiniteDuration,
       zookeeperBlockingExecutionContext: ExecutionContext
-  ): IO[CuratorFramework] =
-//      val _ = Some("foo").asInstanceOf[Option[String]]
-    IO.blocking {
-      curatorFramework(
-        zkQuorumString,
-        zookeeperBlockingExecutionContext,
-        new RetryForever(retryInterval.toMillis.toInt)
-      )
+  ): Resource[IO, CuratorFramework] = {
+    Resource.make {
+      IO.blocking {
+        curatorFramework(
+          zkQuorumString,
+          zookeeperBlockingExecutionContext,
+          new RetryForever(retryInterval.toMillis.toInt)
+        )
+      }
+    } { r =>
+      IO.blocking(r.close())
     }
+  }
 }
